move resolve_credentials to helpers
