+++ qiskit_ionq/ionq_provider.py
@@ -27,51 +27,15 @@
 """Provider for interacting with IonQ backends"""
 
 import logging
-import os
-from dotenv import dotenv_values
 
 from qiskit.providers.exceptions import QiskitBackendNotFoundError
 from qiskit.providers.providerutils import filter_backends
+from .helpers import resolve_credentials
 
 from . import ionq_backend
 
 logger = logging.getLogger(__name__)
 
-
-def resolve_credentials(token: str = None, url: str = None):
-    """Resolve credentials for use in IonQ Client API calls.
-
-    If the provided ``token`` and ``url`` are both ``None``, then these values
-    are loaded from the ``IONQ_API_TOKEN`` and ``IONQ_API_URL``
-    environment variables, respectively.
-
-    If no url is discovered, then ``https://api.ionq.co/v0.3`` is used.
-
-    Args:
-        token (str): IonQ API access token.
-        url (str, optional): IonQ API url. Defaults to ``None``.
-
-    Returns:
-        dict[str]: A dict with "token" and "url" keys, for use by a client.
-    """
-    env_token = (
-        dotenv_values().get("QISKIT_IONQ_API_TOKEN")  # first check for dotenv values
-        or dotenv_values().get("IONQ_API_KEY")
-        or dotenv_values().get("IONQ_API_TOKEN")
-        or os.getenv("QISKIT_IONQ_API_TOKEN")  # then check for global env values
-        or os.getenv("IONQ_API_KEY")
-        or os.getenv("IONQ_API_TOKEN")
-    )
-    env_url = (
-        dotenv_values().get("QISKIT_IONQ_API_URL")
-        or dotenv_values().get("IONQ_API_URL")
-        or os.getenv("QISKIT_IONQ_API_URL")
-        or os.getenv("IONQ_API_URL")
-    )
-    return {
-        "token": token or env_token,
-        "url": url or env_url or "https://api.ionq.co/v0.3",
-    }
 
 
 class IonQProvider:
